Create Cli-ASkpass project.
Split Cli-CredentialHelper into two projects: Cli-CredentialHelper and Cli-Shared.

Moved as much possible shared logic from Cli-CredentialHelper as possible into Cli-Shared.

Created the new Cli-Askpass project and moved the Cli-CredentialHelper::AskPass method into it.
