+++ Microsoft.Alm.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cli-CredentialHelper", "Cli-CredentialHelper\Cli-CredentialHelper.csproj", "{62F52119-63D4-40A8-A9DF-F1C4B473308A}"
 EndProject
@@ -25,7 +25,16 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GitHub.Authentication", "Gi
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GitHub.Authentication.Test", "GitHub.Authentication.Test\GitHub.Authentication.Test.csproj", "{F24E0E80-8DDB-41D4-9F94-DCBC0CE81583}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cli-Askpass", "Cli-AskPass\Cli-Askpass.csproj", "{0653670E-B33A-4EBE-92B8-04C93F43CAE0}"
+EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Cli-Shared", "Cli-Shared\Cli-Shared.shproj", "{DBD202C3-94D5-49D3-8ED1-C72637BA7D90}"
+EndProject
 Global
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		Cli-Shared\Cli-Shared.projitems*{0653670e-b33a-4ebe-92b8-04c93f43cae0}*SharedItemsImports = 4
+		Cli-Shared\Cli-Shared.projitems*{62f52119-63d4-40a8-a9df-f1c4b473308a}*SharedItemsImports = 4
+		Cli-Shared\Cli-Shared.projitems*{dbd202c3-94d5-49d3-8ed1-c72637ba7d90}*SharedItemsImports = 13
+	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
@@ -63,6 +72,10 @@ Global
 		{F24E0E80-8DDB-41D4-9F94-DCBC0CE81583}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F24E0E80-8DDB-41D4-9F94-DCBC0CE81583}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F24E0E80-8DDB-41D4-9F94-DCBC0CE81583}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0653670E-B33A-4EBE-92B8-04C93F43CAE0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0653670E-B33A-4EBE-92B8-04C93F43CAE0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0653670E-B33A-4EBE-92B8-04C93F43CAE0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0653670E-B33A-4EBE-92B8-04C93F43CAE0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
